feat: add API for creating and getting PayButtons
Description: This module provides a simple and mocked API for creating and getting PayButtons.

Test Plan: After clicking Add PayButton on the signed-in index page, it will show on console the response.
Also, after adding a PayButton, a simple list of added buttons with corresponding addresses will be shown on the signed-in page.

Depends on: feat/add-button/2